Added Group Facility/CCOF Funding.

+++ backend/src/routes/facility.js
@@ -23,7 +23,7 @@ router.get(
   '/:facilityId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
-  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION, PERMISSIONS.VIEW_FACILITY_INFORMATION),
+  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION, PERMISSIONS.VIEW_SUBMITTED_PCF, PERMISSIONS.VIEW_FACILITY_INFORMATION),
   validateFacility(),
   [param('facilityId', 'URL param: [facilityId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
@@ -32,6 +32,9 @@ router.get(
   },
 );
 
+/**
+ * Get Facility License Categories
+ */
 router.get(
   '/:facilityId/licenseCategories',
   passport.authenticate('jwt', { session: false }),
@@ -57,7 +60,6 @@ router.get('/ccfri/:ccfriId', passport.authenticate('jwt', { session: false }), 
 
 /**
  * Get Parent Fees for a facility
- *
  */
 // TODO #securitymatrix - Implement with Applications security
 router.get(
@@ -74,8 +76,7 @@ router.get(
 /**
  * Create a new Facility
  */
-// TODO #securitymatrix - Implement with Applications security
-router.post('/', passport.authenticate('jwt', { session: false }), isValidBackendToken, [checkSchema(facilitySchema)], (req, res) => {
+router.post('/', passport.authenticate('jwt', { session: false }), isValidBackendToken, validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION), [checkSchema(facilitySchema)], (req, res) => {
   validationResult(req).throw();
   return createFacility(req, res);
 });
@@ -96,19 +97,19 @@ router.put(
   },
 );
 
-// TODO #securitymatrix - Implement with Applications security
-router.delete('/:facilityId', passport.authenticate('jwt', { session: false }), isValidBackendToken, [param('facilityId', 'URL param: [facilityId] is required').not().isEmpty()], (req, res) => {
-  validationResult(req).throw();
-  return deleteFacility(req, res);
-});
-
 /**
- * Submit a complete application
+ * Delete a Facility
  */
-// TODO #securitymatrix - Implement with Applications security
-router.post('/:facilityId/submit', passport.authenticate('jwt', { session: false }), isValidBackendToken, [checkSchema(facilitySchema)], (req, res) => {
-  validationResult(req).throw();
-  return createFacility(req, res);
-});
+router.delete(
+  '/:facilityId',
+  passport.authenticate('jwt', { session: false }),
+  isValidBackendToken,
+  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION),
+  [param('facilityId', 'URL param: [facilityId] is required').not().isEmpty()],
+  (req, res) => {
+    validationResult(req).throw();
+    return deleteFacility(req, res);
+  },
+);
 
 module.exports = router;
